+++ src/auth/adapter/inbound/controller/AuthLoginRequestHandler.tsx
@@ -3,7 +3,7 @@ import { executeOAuthLoginRequest } from '@/auth/application/usecase/OAuthLoginR
 export async function handleAuthLoginRequest (provider: string): Promise<void> {
   const url = await executeOAuthLoginRequest(provider)
   if (url == null) {
-    console.error('로그인에 실패했습니다: 서버 에러')
+    console.error('로그인 리다이렉트 URL이 없습니다.')
     return
   }
   window.location.href = url.url
