+++ src/auth/application/usecase/OAuthLoginRequestUseCase.ts
@@ -1,4 +1,4 @@
-import { getOAuthLoginUrl } from '@/auth/adapter/api/auth'
+import { getOAuthLoginUrl } from '@/auth/adapter/outbound/api/auth'
 
 export async function executeOAuthLoginRequest (provider: string): Promise<{ url: string } | null> {
   return await getOAuthLoginUrl(provider)
